matrix inversion is here and now the matrices are with double and not int

+++ includes/my_math.h
@@ -3,5 +3,5 @@
 
 #include "my_matrix.h"
 
-int dot_product(int* a, int* b, unsigned int size);
-int my_power(int x, unsigned int n);
+double dot_product(double* a, double* b, unsigned int size);
+double my_power(double x, unsigned int n);
